Move AreaOfInterestType and aoi_to_table into pick_aoi/types.py
